+++ app/src/androidTest/java/woowacourse/movie/uiTest/MainActivityTest.kt
@@ -11,7 +11,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.movie.R
-import woowacourse.movie.activity.MainActivity
+import woowacourse.movie.view.MainActivity
 
 class MainActivityTest {
     @get:Rule
