+++ app/src/main/AndroidManifest.xml
@@ -13,15 +13,15 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".activity.ReservationActivity"
+            android:name=".view.reservation.ReservationActivity"
             android:exported="false"
-            android:parentActivityName=".activity.MainActivity"/>
+            android:parentActivityName=".view.MainActivity"/>
         <activity
-            android:name=".activity.ReservationCompleteActivity"
+            android:name=".view.reservation.ReservationCompleteActivity"
             android:exported="false"
-            android:parentActivityName=".activity.MainActivity" />
+            android:parentActivityName=".view.MainActivity" />
         <activity
-            android:name=".activity.MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
