+++ app/src/main/java/woowacourse/movie/view/adapter/MovieViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.adapter
+package woowacourse.movie.view.adapter
 
 import android.content.Context
 import android.view.View
