+++ app/src/main/java/woowacourse/movie/view/dialog/DialogFactory.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.dialog
+package woowacourse.movie.view.dialog
 
 import android.content.Context
 import androidx.appcompat.app.AlertDialog
