+++ app/src/main/java/woowacourse/movie/view/dialog/DialogInfo.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.dialog
+package woowacourse.movie.view.dialog
 
 import android.content.Context
