+++ app/src/main/java/woowacourse/movie/view/reservation/ReservationCompleteActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.activity
+package woowacourse.movie.view.reservation
 
 import android.content.Context
 import android.content.Intent
@@ -10,8 +10,8 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import woowacourse.movie.R
-import woowacourse.movie.dialog.DialogFactory
 import woowacourse.movie.domain.Ticket
+import woowacourse.movie.view.dialog.DialogFactory
 import java.time.format.DateTimeFormatter
 
 class ReservationCompleteActivity : AppCompatActivity() {
