+++ app/src/main/res/layout/activity_reservation.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activity.ReservationActivity">
+    tools:context=".view.reservation.ReservationActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
